+++ src/main/java/com/depromeet/omobackend/dto/response/UserSaveResponseDto.java
@@ -1,0 +1,16 @@
+package com.depromeet.omobackend.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserSaveResponseDto {
+
+    private String nickname;
+    private String profileUrl;
+}
